Moved some of the terminal-related files into terminal/ folder
