+++ c/utils/terminal/terminal_stream.h
